feat(dataloader): add get_dataset method to access underlying dataset
refactor(core): change ImageShape fields from uint16_t to uint32_t

build(examples): simplify manual_train CMake configuration and rename target

fix(dataloader): remove redundant kernel code and add dataset null check

feat(trainer): implement comprehensive trainer class with training loop

docs(trainer): add detailed documentation for trainer class and components

+++ tinygs/include/tinygs/core/image.hpp
@@ -14,8 +14,8 @@ enum class ImageDataType {
 };
 
 struct ImageShape {
-  uint16_t width, height;
-  uint16_t channel; // 3, 4 is supported.
+  uint32_t width, height;
+  uint32_t channel; // 3, 4 is supported.
 
   TINYGS_HOST_DEVICE bool operator==(const ImageShape& other) const noexcept {
     return width == other.width && height == other.height && channel == other.channel;
